Enhancement: Extract SameSizeTest
