+++ tests/Constraint/ConstraintTest.php
@@ -2636,52 +2636,6 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintSameSizeWithAnArray()
-    {
-        $constraint = new SameSize([1, 2, 3, 4, 5]);
-
-        $this->assertTrue($constraint->evaluate([6, 7, 8, 9, 10], '', true));
-        $this->assertFalse($constraint->evaluate([1, 2, 3, 4], '', true));
-    }
-
-    public function testConstraintSameSizeWithAnIteratorWhichDoesNotImplementCountable()
-    {
-        $constraint = new SameSize(new \TestIterator([1, 2, 3, 4, 5]));
-
-        $this->assertTrue($constraint->evaluate(new \TestIterator([6, 7, 8, 9, 10]), '', true));
-        $this->assertFalse($constraint->evaluate(new \TestIterator([1, 2, 3, 4]), '', true));
-    }
-
-    public function testConstraintSameSizeWithAnObjectImplementingCountable()
-    {
-        $constraint = new SameSize(new \ArrayObject([1, 2, 3, 4, 5]));
-
-        $this->assertTrue($constraint->evaluate(new \ArrayObject([6, 7, 8, 9, 10]), '', true));
-        $this->assertFalse($constraint->evaluate(new \ArrayObject([1, 2, 3, 4]), '', true));
-    }
-
-    public function testConstraintSameSizeFailing()
-    {
-        $constraint = new SameSize([1, 2, 3, 4, 5]);
-
-        try {
-            $constraint->evaluate([1, 2]);
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-Failed asserting that actual size 2 matches expected size 5.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
     public function testConstraintNotSameSizeFailing()
     {
         $constraint = Assert::logicalNot(
